+++ src/pages/FolderPage/FolderPage.jsx
@@ -1,6 +1,6 @@
 import { Helmet, HelmetProvider } from 'react-helmet-async';
-import TopContent from '../components/Folder/TopContent/TopContent';
-import MainContent from '../components/Folder/MainContent/MainContent';
+import TopContent from '../../components/Folder/TopContent/TopContent';
+import MainContent from '../../components/Folder/MainContent/MainContent';
 
 function FolderPage() {
   return (
