+++ src/pages/SharedPage/SharedPage.jsx
@@ -1,6 +1,6 @@
 import { Helmet, HelmetProvider } from 'react-helmet-async';
-import SharedHeader from '../components/Shared/SharedHeader';
-import CardList from '../components/CardList/CardList';
+import SharedHeader from '../../components/Shared/SharedHeader';
+import CardList from '../../components/CardList/CardList';
 
 function SharedPage() {
   return (
